Tidy Up Movement Plugin
- Moved code that should probably reworked into another module
into the orphans.rs file.
