+++ src/main.rs
@@ -336,32 +336,32 @@ fn game_over_on_players_death(
 //     mut enemies_query: Query<(Entity, &mut TripPointX), (With<Enemy>, Without<Target>)>,
 //     player_query: Query<&Transform, With<Player>>,
 // ) {
-    // let mut rng = rand::thread_rng();
-    // let p_transforms = player_query.iter().collect::<Vec<_>>();
-    // let max_player_x = p_transforms
-    //     .iter()
-    //     .map(|transform| transform.translation.x)
-    //     .max_by(f32::total_cmp);
+// let mut rng = rand::thread_rng();
+// let p_transforms = player_query.iter().collect::<Vec<_>>();
+// let max_player_x = p_transforms
+//     .iter()
+//     .map(|transform| transform.translation.x)
+//     .max_by(f32::total_cmp);
 
-    // if let Some(max_player_x) = max_player_x {
-    //     for (e_entity, e_state, mut e_trip_point_x) in enemies_query.iter_mut() {
-    //         if *e_state == State::Idle {
-    //             if let Some(p_transform) = p_transforms.choose(&mut rng) {
-    //                 if max_player_x > e_trip_point_x.0 {
-    //                     e_trip_point_x.0 = f32::MIN;
+// if let Some(max_player_x) = max_player_x {
+//     for (e_entity, e_state, mut e_trip_point_x) in enemies_query.iter_mut() {
+//         if *e_state == State::Idle {
+//             if let Some(p_transform) = p_transforms.choose(&mut rng) {
+//                 if max_player_x > e_trip_point_x.0 {
+//                     e_trip_point_x.0 = f32::MIN;
 
-    //                     let x_offset = rng.gen_range(-100.0..100.);
-    //                     let y_offset = rng.gen_range(-100.0..100.);
-    //                     commands.entity(e_entity).insert(Target {
-    //                         position: Vec2::new(
-    //                             p_transform.translation.x + x_offset,
-    //                             (p_transform.translation.y + y_offset)
-    //                                 .clamp(consts::MIN_Y, consts::MAX_Y),
-    //                         ),
-    //                     });
-    //                 }
-    //             }
-    //         }
-    //     }
-    // }
+//                     let x_offset = rng.gen_range(-100.0..100.);
+//                     let y_offset = rng.gen_range(-100.0..100.);
+//                     commands.entity(e_entity).insert(Target {
+//                         position: Vec2::new(
+//                             p_transform.translation.x + x_offset,
+//                             (p_transform.translation.y + y_offset)
+//                                 .clamp(consts::MIN_Y, consts::MAX_Y),
+//                         ),
+//                     });
+//                 }
+//             }
+//         }
+//     }
+// }
 // }
